8032733: Fix cast lint warnings in client libraries 8032734: Add cast lint warning to build of jdk repository

+++ jdk/src/share/classes/sun/font/StrikeCache.java
@@ -351,7 +351,7 @@ public final class StrikeCache {
                         if (gids == null) {
                             gids = new ArrayList<Long>();
                         }
-                        gids.add((long) glyphPtrs[i]);
+                        gids.add(glyphPtrs[i]);
                     }
                 }
